+++ react/modules/debouncedPixelHook.ts
@@ -1,10 +1,10 @@
 import { useRef } from 'react'
-import { usePixel } from 'vtex.pixel-manager/PixelContext'
+import { usePixel } from 'vtex.pixel-manager'
 import debounce from 'debounce'
 
 export default function useDebouncedPush() {
   const { push } = usePixel()
-  const debouncedPush = useRef<(param: object) => void>(debounce(push, 600))
+  const debouncedPush = useRef<typeof push>(debounce(push, 600))
 
   return debouncedPush.current
 }